Refatora QUICK_ATTACK de Warrior para Multiclasse
